Move tests to a new home
