+++ source/Energinet.DataHub.MarketRoles.sln
@@ -21,8 +21,6 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Processing.Domain", "Proces
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Energinet.DataHub.MarketRoles.ApplyDBMigrationsApp", "Energinet.DataHub.MarketRoles.ApplyDBMigrationsApp\Energinet.DataHub.MarketRoles.ApplyDBMigrationsApp.csproj", "{2B806039-5BCF-4768-95FE-56778403E64D}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Messaging.CimMessageAdapter.Tests", "Messaging.CimMessageAdapter.Tests\Messaging.CimMessageAdapter.Tests.csproj", "{0B8E940E-A2CE-458B-BB51-C12050D91360}"
-EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Messaging.CimMessageAdapter", "Messaging.CimMessageAdapter\Messaging.CimMessageAdapter.csproj", "{27395443-B503-4A4A-B9F1-01C3F9C75222}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "B2B", "B2B", "{1C877C68-8E3D-4371-B96A-7360850A538E}"
@@ -87,10 +85,6 @@ Global
 		{2B806039-5BCF-4768-95FE-56778403E64D}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{2B806039-5BCF-4768-95FE-56778403E64D}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{2B806039-5BCF-4768-95FE-56778403E64D}.Release|Any CPU.Build.0 = Release|Any CPU
-		{0B8E940E-A2CE-458B-BB51-C12050D91360}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{0B8E940E-A2CE-458B-BB51-C12050D91360}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{0B8E940E-A2CE-458B-BB51-C12050D91360}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{0B8E940E-A2CE-458B-BB51-C12050D91360}.Release|Any CPU.Build.0 = Release|Any CPU
 		{27395443-B503-4A4A-B9F1-01C3F9C75222}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{27395443-B503-4A4A-B9F1-01C3F9C75222}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{27395443-B503-4A4A-B9F1-01C3F9C75222}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -147,7 +141,6 @@ Global
 		{79DE71D0-7AA6-4B96-AB44-85F4C621608E} = {582BFE4E-BBFB-4CBE-840D-2F97C2BA1D3C}
 		{3801DF77-3A40-460C-9529-7AD9422037B0} = {582BFE4E-BBFB-4CBE-840D-2F97C2BA1D3C}
 		{87D55C95-0790-42A1-B543-AF27BF2E0950} = {582BFE4E-BBFB-4CBE-840D-2F97C2BA1D3C}
-		{0B8E940E-A2CE-458B-BB51-C12050D91360} = {1C877C68-8E3D-4371-B96A-7360850A538E}
 		{27395443-B503-4A4A-B9F1-01C3F9C75222} = {1C877C68-8E3D-4371-B96A-7360850A538E}
 		{B14EBA74-CCAF-4677-8AE5-A0F6302ED54C} = {1C877C68-8E3D-4371-B96A-7360850A538E}
 		{76FA7684-27E1-4022-9FB3-F72811E45C60} = {1C877C68-8E3D-4371-B96A-7360850A538E}
