--- source/Messaging.CimMessageAdapter.Tests/Messaging.CimMessageAdapter.Tests.csproj
@@ -1,65 +1,0 @@
-<!--
-Copyright 2020 Energinet DataHub A/S
-
-Licensed under the Apache License, Version 2.0 (the "License2");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
--->
-<Project Sdk="Microsoft.NET.Sdk">
-
-    <PropertyGroup>
-        <TargetFramework>net6.0</TargetFramework>
-
-        <IsPackable>false</IsPackable>
-    </PropertyGroup>
-
-    <ItemGroup>
-        <PackageReference Include="Energinet.DataHub.Core.Schemas" Version="1.0.9" />
-        <PackageReference Include="Microsoft.NET.Test.Sdk" Version="17.1.0" />
-        <PackageReference Include="xunit" Version="2.4.1" />
-        <PackageReference Include="xunit.categories" Version="2.0.6" />
-        <PackageReference Include="xunit.runner.visualstudio" Version="2.4.3">
-            <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
-            <PrivateAssets>all</PrivateAssets>
-        </PackageReference>
-        <PackageReference Include="coverlet.collector" Version="3.1.2">
-            <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
-            <PrivateAssets>all</PrivateAssets>
-        </PackageReference>
-    </ItemGroup>
-
-    <ItemGroup>
-      <None Remove="ValidMessage.xml" />
-      <None Remove="Schema.xsd" />
-      <None Remove="InvalidMessageContainingTwoErrors.xml" />
-      <None Update="RequestChangeOfSupplier.xml">
-        <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-      </None>
-      <None Update="RequestChangeOfSupplier_postman.xml">
-        <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-      </None>
-    </ItemGroup>
-
-    <ItemGroup>
-      <None Remove="InvalidRequestChangeOfSupplier.xml" />
-      <None Remove="RequestChangeOfSupplierWithDuplicateTransactionIds.xml" />
-      <None Remove="urn-ediel-org-structure-requestchangeofsupplier-0-1.xsd" />
-      <None Remove="urn-entsoe-eu-local-extension-types.xsd" />
-      <None Remove="urn-entsoe-eu-wgedi-codelists.xsd" />
-      <None Remove="ValidRequestChangeOfSupplier.xml" />
-    </ItemGroup>
-
-    <ItemGroup>
-      <ProjectReference Include="..\Messaging.CimMessageAdapter\Messaging.CimMessageAdapter.csproj" />
-      <ProjectReference Include="..\Messaging.Infrastructure\Messaging.Infrastructure.csproj" />
-    </ItemGroup>
-
-</Project>
