+++ source/Messaging.Tests/Messaging.Tests.csproj
@@ -26,6 +26,7 @@ limitations under the License.
         <PackageReference Include="Microsoft.NET.Test.Sdk" Version="16.9.4" />
         <PackageReference Include="NJsonSchema" Version="10.7.2" />
         <PackageReference Include="xunit" Version="2.4.1" />
+        <PackageReference Include="xunit.categories" Version="2.0.6" />
         <PackageReference Include="xunit.runner.visualstudio" Version="2.4.3">
             <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
             <PrivateAssets>all</PrivateAssets>
